+++ modules/scfmDiagnostics/NEWS.md
@@ -3,4 +3,5 @@
 - completed conversion to using `terra` instead of `raster` objects;
 - completed conversion to using `sf` instead of `sp` objects;
 - updated objects `fireRegimePolys` replaces objects `landscapeAttr`, `scfmDriverPars` and `scfmRegimePars`, which have been removed;
+- added `runName` from SpaDES as a title for `scfmDiagnostics` plots